Add files via upload
It's sort of messy since I was still messing around with it but the base stuff is there. It has a model and view, along with a database configuration class that (hopefully) will help it connect to the insulin pump database. The controller is blank at the moment- I took out code that I could not get to work correctly.